LayoutTests/http/tests/websocket/tests/hybi/websocket-blocked-sending-cookie-as-third-party-from-worker-expected.txt

Tests that a DedicatedWorker WebSocket does not send third-party cookies when ITP cookie blocking is enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS  is undefined.
PASS document.location.host is "127.0.0.1:8000"

Sending third-party cookie through cross-origin WebSocket handshake from DedicatedWorker is blocked.
Created a socket to 'ws://localhost:8880/websocket/tests/hybi/websocket-blocked-sending-cookie-as-third-party'; readyState 0.
PASS Connection was allowed (request did not contain cookies).

PASS successfullyParsed is true

TEST COMPLETE
